requirements.txt
requests~=2.26.0
Flask~=2.0.2
pika~=1.2.0
pytest~=6.2.5
PyMySQL~=1.0.2
PyYAML~=5.4.1
remi~=2021.3.2
PySimpleGUI~=4.49.0